requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
jinja2==3.1.3
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-multipart==0.0.6
google-genai==1.62.0
git+https://github.com/ooda-AI-GB/viv-auth.git
git+https://github.com/ooda-AI-GB/viv-pay.git
